Moving xml_to_dict to siepic_tools.utils.tech
