+++ klayout_dot_config/python/SiEPIC/utils/__init__.py
@@ -35,8 +35,7 @@ translate_from_normal
 pt_intersects_segment
 layout_pgtext
 find_automated_measurement_labels
-etree_to_dict: XML parser
-xml_to_dict
+xml_to_dict: XML parser
 eng_str
 svg_from_component
 sample_function
@@ -127,30 +126,9 @@ def get_technology_by_name(tech_name, verbose=False):
         technology['INTC_CML_version'] = ''
 
     # Layers:
-    file = open(lyp_file, 'r')
-    layer_dict = xml_to_dict(file.read())['layer-properties']['properties']
-    file.close()
-
-    for k in layer_dict:
-        layerInfo = k['source'].split('@')[0]
-        if 'group-members' in k:
-            # encoutered a layer group, look inside:
-            j = k['group-members']
-            if 'name' in j:
-                layerInfo_j = j['source'].split('@')[0]
-                technology[j['name']] = pya.LayerInfo(
-                    int(layerInfo_j.split('/')[0]), int(layerInfo_j.split('/')[1]))
-            else:
-                for j in k['group-members']:
-                    layerInfo_j = j['source'].split('@')[0]
-                    technology[j['name']] = pya.LayerInfo(
-                        int(layerInfo_j.split('/')[0]), int(layerInfo_j.split('/')[1]))
-            if k['source'] != '*/*@*':
-                technology[k['name']] = pya.LayerInfo(
-                    int(layerInfo.split('/')[0]), int(layerInfo.split('/')[1]))
-        else:
-            technology[k['name']] = pya.LayerInfo(
-                int(layerInfo.split('/')[0]), int(layerInfo.split('/')[1]))
+    from siepic_tools.utils.tech import parse_layer_map
+    layer_map = parse_layer_map(lyp_file)
+    technology.update(layer_map)
 
     return technology
 # end of get_technology_by_name(tech_name)
@@ -882,37 +860,7 @@ except NameError:
         return it.next()
 
 
-# XML to Dict parser, from:
-# https://stackoverflow.com/questions/2148119/how-to-convert-an-xml-string-to-a-dictionary-in-python/10077069
-def etree_to_dict(t):
-    from collections import defaultdict
-    d = {t.tag: {} if t.attrib else None}
-    children = list(t)
-    if children:
-        dd = defaultdict(list)
-        for dc in map(etree_to_dict, children):
-            for k, v in dc.items():
-                dd[k].append(v)
-        d = {t.tag: {k: v[0] if len(v) == 1 else v for k, v in dd.items()}}
-    if t.attrib:
-        d[t.tag].update(('@' + k, v) for k, v in t.attrib.items())
-    if t.text:
-        text = t.text.strip()
-        if children or t.attrib:
-            if text:
-                d[t.tag]['#text'] = text
-        else:
-            d[t.tag] = text
-    return d
-
-
-def xml_to_dict(t):
-    from xml.etree import cElementTree as ET
-    try:
-        e = ET.XML(t)
-    except:
-        raise UserWarning("Error in the XML file.")
-    return etree_to_dict(e)
+from siepic_tools.utils.tech import xml_to_dict
 
 
 def eng_str(x):
